Removed AbstrasctProcessingContainer,AbstractServiceContainer abstractions and renamed DefaultProcessingContainer to ProcessingContainer
