--- hazelcast-jet-core/src/main/java/com/hazelcast/jet/impl/container/AbstractServiceContainer.java
@@ -1,33 +1,0 @@
-/*
- * Copyright (c) 2008-2016, Hazelcast, Inc. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.hazelcast.jet.impl.container;
-
-import com.hazelcast.jet.impl.application.ApplicationContext;
-import com.hazelcast.jet.impl.statemachine.StateMachine;
-import com.hazelcast.jet.impl.statemachine.StateMachineFactory;
-import com.hazelcast.spi.NodeEngine;
-
-public abstract class AbstractServiceContainer
-        <SI extends ContainerEvent,
-                SS extends ContainerState,
-                SO extends ContainerResponse> extends AbstractContainer<SI, SS, SO> {
-    public AbstractServiceContainer(StateMachineFactory<SI, StateMachine<SI, SS, SO>> stateMachineFactory,
-                                    NodeEngine nodeEngine,
-                                    ApplicationContext applicationContext) {
-        super(stateMachineFactory, nodeEngine, applicationContext, null);
-    }
-}
