--- hazelcast-jet-core/src/main/java/com/hazelcast/jet/impl/container/DefaultProcessingContainer.java
@@ -1,38 +1,0 @@
-/*
- * Copyright (c) 2008-2016, Hazelcast, Inc. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.hazelcast.jet.impl.container;
-
-import com.hazelcast.jet.dag.Vertex;
-import com.hazelcast.jet.data.tuple.JetTupleFactory;
-import com.hazelcast.jet.impl.application.ApplicationContext;
-import com.hazelcast.jet.processor.ContainerProcessor;
-import com.hazelcast.spi.NodeEngine;
-
-import java.util.function.Supplier;
-
-public class DefaultProcessingContainer extends AbstractProcessingContainer {
-    public DefaultProcessingContainer(Vertex vertex,
-                                      Supplier<ContainerProcessor> containerProcessorFactory,
-                                      NodeEngine nodeEngine,
-                                      ApplicationContext applicationContext, JetTupleFactory tupleFactory) {
-        super(vertex, containerProcessorFactory, nodeEngine, applicationContext, tupleFactory);
-    }
-
-    protected void wakeUpExecutor() {
-        getApplicationContext().getExecutorContext().getDataContainerStateMachineExecutor().wakeUp();
-    }
-}
