+++ hazelcast-jet-core/src/main/java/com/hazelcast/jet/impl/container/events/AbstractEventProcessor.java
@@ -27,7 +27,6 @@ import com.hazelcast.jet.impl.container.ContainerTask;
 import com.hazelcast.jet.impl.container.ProcessingContainer;
 import com.hazelcast.jet.impl.container.processingcontainer.ProcessingContainerEvent;
 import com.hazelcast.logging.ILogger;
-
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
