+++ hazelcast-jet-core/src/main/java/com/hazelcast/jet/impl/container/events/DefaultEventProcessorFactory.java
@@ -20,7 +20,6 @@ import com.hazelcast.jet.impl.container.ContainerContext;
 import com.hazelcast.jet.impl.container.ContainerTask;
 import com.hazelcast.jet.impl.container.ProcessingContainer;
 import com.hazelcast.jet.impl.container.task.TaskEvent;
-
 import java.util.IdentityHashMap;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicInteger;
