+++ hazelcast-jet-core/src/main/java/com/hazelcast/jet/impl/container/events/TaskEventExecutionErrorProcessor.java
@@ -21,7 +21,6 @@ import com.hazelcast.jet.impl.container.ContainerContext;
 import com.hazelcast.jet.impl.container.ContainerTask;
 import com.hazelcast.jet.impl.container.ProcessingContainer;
 import com.hazelcast.jet.impl.container.task.TaskEvent;
-
 import java.util.concurrent.atomic.AtomicInteger;
 
 public class TaskEventExecutionErrorProcessor extends AbstractEventProcessor {
