+++ hazelcast-jet-core/src/main/java/com/hazelcast/jet/impl/container/task/DefaultContainerTask.java
@@ -43,7 +43,6 @@ import com.hazelcast.jet.processor.ContainerProcessor;
 import com.hazelcast.logging.ILogger;
 import com.hazelcast.nio.Address;
 import com.hazelcast.spi.NodeEngine;
-
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
