+++ hazelcast-jet-core/src/main/java/com/hazelcast/jet/impl/container/task/nio/DefaultSocketReader.java
@@ -31,7 +31,6 @@ import com.hazelcast.jet.impl.executor.Payload;
 import com.hazelcast.jet.impl.util.JetUtil;
 import com.hazelcast.nio.Address;
 import com.hazelcast.spi.NodeEngine;
-
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.nio.channels.SocketChannel;
