+++ hazelcast-jet-core/src/main/java/com/hazelcast/jet/impl/container/task/processors/shuffling/ShuffledActorTaskProcessor.java
@@ -30,7 +30,6 @@ import com.hazelcast.jet.impl.data.io.DefaultObjectIOStream;
 import com.hazelcast.jet.impl.util.JetUtil;
 import com.hazelcast.jet.processor.ContainerProcessor;
 import com.hazelcast.nio.Address;
-
 import java.util.ArrayList;
 import java.util.List;
