+++ hazelcast-jet-core/src/main/java/com/hazelcast/jet/impl/container/task/processors/shuffling/ShuffledConsumerTaskProcessor.java
@@ -38,7 +38,6 @@ import com.hazelcast.jet.strategy.ShufflingStrategy;
 import com.hazelcast.nio.Address;
 import com.hazelcast.spi.NodeEngine;
 import com.hazelcast.util.HashUtil;
-
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
