+++ hazelcast-jet-core/src/main/java/com/hazelcast/jet/impl/statemachine/applicationmaster/processors/ExecuteApplicationProcessor.java
@@ -27,7 +27,6 @@ import com.hazelcast.jet.impl.container.ProcessingContainer;
 import com.hazelcast.jet.impl.executor.ApplicationTaskContext;
 import com.hazelcast.jet.impl.statemachine.container.requests.ContainerExecuteRequest;
 import com.hazelcast.logging.ILogger;
-
 import java.util.Iterator;
 import java.util.concurrent.TimeUnit;
