+++ hazelcast-jet-core/src/main/java/com/hazelcast/jet/impl/statemachine/applicationmaster/processors/InterruptApplicationProcessor.java
@@ -21,7 +21,6 @@ import com.hazelcast.jet.impl.container.ApplicationMaster;
 import com.hazelcast.jet.impl.container.ContainerPayloadProcessor;
 import com.hazelcast.jet.impl.container.ProcessingContainer;
 import com.hazelcast.jet.impl.statemachine.container.requests.ContainerInterruptRequest;
-
 import java.util.concurrent.TimeUnit;
 
 public class InterruptApplicationProcessor implements ContainerPayloadProcessor<Dummy> {
